extracted yesno prompt from compare&cli, rm unused quit() from compare
